refactor(product): extract product view and route parsing
